Example 8.1.2 from Experimental Design & Analysis for Tree Improvement

Exam8.1.2 presents the Analysis of Nested Seedlot Structure of Diameter at breast height (Dbh) of 60 SeedLots under layout of row column design with 6 rows and 10 columns in 18 countries and 59 provinces of 18 selected countries given in Example 8.1.

References

Williams, E.R., Matheson, A.C. and Harwood, C.E. (2002).Experimental Design and Analysis for Tree Improvement. CSIRO Publishing.

See also

Examples

data(DataExam8.1)
library(tidyverse)
library(emmeans)
  fm8.11 <- aov(formula      =
  Dbh~Country*Province
  ,data         = DataExam8.1
  #, subset
  #, weights
  #, na.action
  , method      = "qr"
  , model       = TRUE
  , x           = FALSE
  , y           = FALSE
  , qr          = TRUE
  , singular.ok = TRUE
  , contrasts   = NULL
  )
  b<-anova(fm8.11)
     Res                     <- length(b[["Sum Sq"]])
     df                      <- 119
     MSS                     <- 0.1951
     b[["Df"]][Res]          <- df
     b[["Sum Sq"]][Res]      <- MSS*df
     b[["Mean Sq"]][Res]     <- b[["Sum Sq"]][Res]/b[["Df"]][Res]
     b[["F value"]][1:Res-1] <- b[["Mean Sq"]][1:Res-1]/b[["Mean Sq"]][Res]
     b[["Pr(>F)"]][Res-1]     <- df(b[["F value"]][Res-1],b[["Df"]][Res-1],b[["Df"]][Res])
     b

#> Analysis of Variance Table
#> 
#> Response: Dbh
#>            Df Sum Sq Mean Sq F value    Pr(>F)    
#> Country    17 54.651  3.2148 16.4777 1.177e-08 ***
#> Province   41 18.653  0.4549  2.3318  0.001442 ** 
#> Residuals 119 23.217  0.1951                      
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

  library(emmeans)
  emmeans::emmeans(fm8.11,specs = "Country")

#> NOTE: A nesting structure was detected in the fitted model:
#>     Province %in% Country
#> If this is incorrect, re-run or update with `nesting` specified

#> NOTE: Results may be misleading due to involvement in interactions

#>  Country       emmean        SE  df lower.CL upper.CL
#>  Australia   2.631250 0.2016697 177 2.233264 3.029236
#>  Benin       3.342500 0.4033393 177 2.546527 4.138473
#>  China       3.685833 0.2328681 177 3.226278 4.145388
#>  Egypt       2.498333 0.2328681 177 2.038778 2.957888
#>  Fiji        2.611667 0.2328681 177 2.152112 3.071222
#>  Guam        2.342500 0.4033393 177 1.546527 3.138473
#>  India       3.574583 0.1646626 177 3.249629 3.899538
#>  Kenya       3.491250 0.1426020 177 3.209831 3.772669
#>  Malaysia    4.030833 0.1344464 177 3.765509 4.296158
#>  Mauritius   3.122500 0.4033393 177 2.326527 3.918473
#>  PNG         3.650000 0.4033393 177 2.854027 4.445973
#>  Philippines 3.611667 0.2328681 177 3.152112 4.071222
#>  Puerto Rico 3.345000 0.4033393 177 2.549027 4.140973
#>  Solomon IS  3.698750 0.2852040 177 3.135912 4.261588
#>  Sri Lanks   3.235000 0.2328681 177 2.775445 3.694555
#>  Thailand    3.841250 0.2016697 177 3.443264 4.239236
#>  Vanuatu     2.762500 0.4033393 177 1.966527 3.558473
#>  Vietnam     3.276500 0.1803788 177 2.920530 3.632470
#> 
#> Results are averaged over the levels of: Province 
#> Confidence level used: 0.95