Example 8.1.1 from Experimental Design & Analysis for Tree Improvement

Exam8.1.1 presents the Mixed Effects Analysis of Diameter at breast height (Dbh) of 60 SeedLots under layout of row column design with 6 rows and 10 columns in 18 countries and 59 provinces of 18 selected countries given in Example 8.1.

References

Williams, E.R., Matheson, A.C. and Harwood, C.E. (2002).Experimental Design and Analysis for Tree Improvement. CSIRO Publishing.

See also

Examples

data(DataExam8.1)
library(tidyverse)
library(emmeans)

#> 
#> Attaching package: ‘emmeans’

#> The following object is masked from ‘package:devtools’:
#> 
#>     test

library(lme4)
fm8.8<- lmer(formula=
     Dbh~1+Replication+Column+Province+(1|Replication:Row)+(1|Replication:Column)
    ,data = DataExam8.1
    ,REML = TRUE
    ,control = lmerControl()
    ,start = NULL
    , verbose = 0L
   #, subset
   #, weights
   #, na.action
   #,offset
    , contrasts = NULL
    , devFunOnly = FALSE)
  anova(fm8.8)

#> Type III Analysis of Variance Table with Satterthwaite's method
#>             Sum Sq Mean Sq NumDF   DenDF F value    Pr(>F)    
#> Replication  2.695 0.89819     3  21.422  4.5587   0.01281 *  
#> Column      24.882 2.76470     9  23.702 14.0321 1.652e-07 ***
#> Province    55.316 0.95372    58 137.058  4.8406 1.916e-14 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

  summary(fm8.8)

#> Linear mixed model fit by REML. t-tests use Satterthwaite's method [
#> lmerModLmerTest]
#> Formula: Dbh ~ 1 + Replication + Column + Province + (1 | Replication:Row) +  
#>     (1 | Replication:Column)
#>    Data: DataExam8.1
#> 
#> REML criterion at convergence: 358.6
#> 
#> Scaled residuals: 
#>      Min       1Q   Median       3Q      Max 
#> -2.69863 -0.50295  0.04918  0.54501  1.67530 
#> 
#> Random effects:
#>  Groups             Name        Variance Std.Dev.
#>  Replication:Column (Intercept) 0.04607  0.2146  
#>  Replication:Row    (Intercept) 0.05987  0.2447  
#>  Residual                       0.19703  0.4439  
#> Number of obs: 236, groups:  Replication:Column, 40; Replication:Row, 24
#> 
#> Fixed effects:
#>               Estimate Std. Error        df t value Pr(>|t|)    
#> (Intercept)    2.19693    0.31958 117.21304   6.874 3.21e-10 ***
#> Replication1   0.25290    0.18945  21.42648   1.335 0.195907    
#> Replication2  -0.37215    0.18945  21.42615  -1.964 0.062587 .  
#> Replication3  -0.29173    0.18944  21.42572  -1.540 0.138205    
#> Column1        1.57267    0.20944  23.92708   7.509 9.69e-08 ***
#> Column2        1.57611    0.21140  24.73585   7.456 8.86e-08 ***
#> Column3        1.92846    0.20872  23.49317   9.239 2.74e-09 ***
#> Column4        1.89916    0.21162  24.87311   8.974 2.84e-09 ***
#> Column5        1.66722    0.20932  23.77696   7.965 3.65e-08 ***
#> Column6        1.85837    0.21406  25.93299   8.682 3.77e-09 ***
#> Column7        1.37220    0.20864  23.55610   6.577 9.21e-07 ***
#> Column8        1.58176    0.21001  24.09210   7.532 8.80e-08 ***
#> Column9        1.24688    0.20969  23.92387   5.946 3.94e-06 ***
#> Province1     -1.12086    0.35735 154.53827  -3.137 0.002047 ** 
#> Province2     -0.41064    0.35091 154.45825  -1.170 0.243718    
#> Province3     -0.69195    0.35421 154.33820  -1.953 0.052570 .  
#> Province4     -1.29899    0.35083 150.33646  -3.703 0.000299 ***
#> Province5      0.17844    0.34835 149.17459   0.512 0.609227    
#> Province6     -0.02596    0.34766 148.58375  -0.075 0.940575    
#> Province7      0.37005    0.35561 157.41215   1.041 0.299652    
#> Province8     -0.84757    0.35539 156.70248  -2.385 0.018279 *  
#> Province10    -0.88853    0.35717 153.46108  -2.488 0.013926 *  
#> Province11    -1.52576    0.35331 154.63600  -4.318 2.80e-05 ***
#> Province12    -0.79366    0.35739 156.42296  -2.221 0.027810 *  
#> Province13    -0.60219    0.35151 154.55947  -1.713 0.088695 .  
#> Province14    -0.90698    0.34931 151.37705  -2.596 0.010347 *  
#> Province15    -0.11556    0.34319 145.71860  -0.337 0.736819    
#> Province16     0.30314    0.35043 153.31750   0.865 0.388360    
#> Province17    -0.11992    0.34866 150.15821  -0.344 0.731374    
#> Province18    -0.20348    0.34912 151.06169  -0.583 0.560872    
#> Province19     0.09508    0.35538 154.14076   0.268 0.789403    
#> Province20    -0.19568    0.34382 147.41694  -0.569 0.570140    
#> Province21    -0.32921    0.34948 151.45568  -0.942 0.347695    
#> Province22    -0.23595    0.35221 150.95625  -0.670 0.503927    
#> Province23    -0.31275    0.34417 146.82385  -0.909 0.365003    
#> Province24     0.13075    0.35529 153.65008   0.368 0.713367    
#> Province25     0.27888    0.34821 152.39529   0.801 0.424440    
#> Province26    -0.03421    0.35352 154.95324  -0.097 0.923024    
#> Province27    -0.06068    0.35430 155.80622  -0.171 0.864232    
#> Province28    -0.45935    0.34505 145.58738  -1.331 0.185181    
#> Province29     0.34293    0.35097 151.58973   0.977 0.330089    
#> Province30    -0.01842    0.34283 147.90853  -0.054 0.957234    
#> Province31     0.55346    0.33729 139.92467   1.641 0.103060    
#> Province32     0.72292    0.35515 156.14791   2.036 0.043488 *  
#> Province33     0.15425    0.34248 147.07132   0.450 0.653094    
#> Province34     0.32147    0.35687 155.74988   0.901 0.369083    
#> Province35     0.52786    0.35287 155.74380   1.496 0.136698    
#> Province36     0.29020    0.33781 140.07119   0.859 0.391768    
#> Province37    -0.13355    0.35416 155.74062  -0.377 0.706613    
#> Province38    -0.49737    0.35522 154.03904  -1.400 0.163477    
#> Province39    -0.05274    0.35110 149.70613  -0.150 0.880799    
#> Province40     0.17266    0.33664 138.56291   0.513 0.608847    
#> Province41     0.19330    0.35472 155.41907   0.545 0.586566    
#> Province42     0.31487    0.34986 151.69504   0.900 0.369559    
#> Province45     0.64922    0.35654 153.87195   1.821 0.070567 .  
#> Province46    -0.08428    0.34899 153.31145  -0.241 0.809492    
#> Province47     0.89895    0.34659 146.61101   2.594 0.010459 *  
#> Province48     0.22176    0.35701 153.25391   0.621 0.535417    
#> Province50    -1.02933    0.34405 147.39872  -2.992 0.003252 ** 
#> Province51    -0.10233    0.35614 157.37349  -0.287 0.774233    
#> Province52    -0.21689    0.35752 154.47563  -0.607 0.544963    
#> Province53     0.02792    0.34993 151.24031   0.080 0.936515    
#> Province54    -0.03048    0.35069 150.46682  -0.087 0.930849    
#> Province55    -0.82604    0.34750 149.08183  -2.377 0.018718 *  
#> Province56    -0.80252    0.35760 159.29520  -2.244 0.026196 *  
#> Province57     0.41416    0.34465 148.24441   1.202 0.231407    
#> Province58    -0.57308    0.34933 151.99835  -1.641 0.102961    
#> Province59    -0.15991    0.34991 150.70782  -0.457 0.648324    
#> Province60    -0.55456    0.35511 156.64863  -1.562 0.120388    
#> Province61    -0.37363    0.35135 152.46814  -1.063 0.289281    
#> Province62    -0.30828    0.34095 140.16835  -0.904 0.367448    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

#> 
#> Correlation matrix not shown by default, as p = 71 > 12.
#> Use print(x, correlation=TRUE)  or
#> 	 vcov(x)	 if you need it

  anova(fm8.8, ddf = "Kenward-Roger")

#> Type III Analysis of Variance Table with Kenward-Roger's method
#>             Sum Sq Mean Sq NumDF   DenDF F value    Pr(>F)    
#> Replication  2.694 0.89811     3  22.453  4.5583   0.01226 *  
#> Column      24.809 2.75650     9  22.986 13.9898 2.268e-07 ***
#> Province    54.691 0.94295    58 134.064  4.7842 4.195e-14 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

  emmeans::emmeans(fm8.8,specs ="Replication")

#>  Replication   emmean        SE    df lower.CL upper.CL
#>  1           3.754334 0.1339419 22.45 3.476876 4.031791
#>  2           3.129287 0.1339432 22.45 2.851827 3.406746
#>  3           3.209701 0.1339369 22.45 2.932253 3.487148
#>  4           3.501433 0.1339879 22.47 3.223893 3.778974
#> 
#> Results are averaged over the levels of: Column, Province 
#> Degrees-of-freedom method: kenward-roger 
#> Confidence level used: 0.95 

  emmeans::emmeans(fm8.8,specs ="Column")

#>  Column   emmean        SE    df lower.CL upper.CL
#>  1      3.501079 0.1550087 26.78 3.182905 3.819253
#>  2      3.504513 0.1564558 27.80 3.183926 3.825101
#>  3      3.856865 0.1548582 26.74 3.538979 4.174751
#>  4      3.827564 0.1549569 26.79 3.509504 4.145624
#>  5      3.595628 0.1549173 26.76 3.277633 3.913622
#>  6      3.786775 0.1573353 28.23 3.464606 4.108943
#>  7      3.300607 0.1548100 26.71 2.982802 3.618411
#>  8      3.510167 0.1569450 28.04 3.188699 3.831635
#>  9      3.175282 0.1548719 26.75 2.857371 3.493194
#>  10     1.928407 0.1572168 28.25 1.606492 2.250321
#> 
#> Results are averaged over the levels of: Replication, Province 
#> Degrees-of-freedom method: kenward-roger 
#> Confidence level used: 0.95 

  emmeans::emmeans(fm8.8,specs ="Province")

#>  Province   emmean        SE     df lower.CL upper.CL
#>  1        2.443605 0.2572392 162.87 1.935651 2.951559
#>  2        3.153833 0.2590290 163.59 2.642362 3.665305
#>  3        2.872521 0.2561955 162.14 2.366611 3.378431
#>  4        2.265480 0.2583426 163.44 1.755361 2.775600
#>  5        3.742912 0.2568539 162.34 3.235707 4.250118
#>  6        3.538508 0.2599454 164.54 3.025249 4.051766
#>  7        3.934518 0.2578819 163.38 3.425307 4.443729
#>  8        2.716899 0.2566741 162.67 2.210057 3.223742
#>  10       2.675941 0.2569938 162.56 2.168465 3.183418
#>  11       2.038708 0.2562404 162.80 1.532724 2.544691
#>  12       2.770811 0.2571088 162.50 2.263106 3.278516
#>  13       2.962280 0.2575629 163.21 2.453695 3.470865
#>  14       2.657488 0.2561172 161.82 2.151725 3.163250
#>  15       3.448912 0.2558282 161.61 2.943715 3.954109
#>  16       3.867612 0.2557240 162.31 3.362637 4.372587
#>  17       3.444553 0.2566187 162.56 2.937817 3.951289
#>  18       3.360986 0.2571419 162.85 2.853224 3.868748
#>  19       3.659551 0.2568099 162.48 3.152436 4.166667
#>  20       3.368791 0.2567583 162.54 2.861779 3.875803
#>  21       3.235259 0.2554642 161.38 2.730775 3.739743
#>  22       3.328514 0.2573288 163.16 2.820390 3.836638
#>  23       3.251721 0.2576640 163.33 2.742939 3.760503
#>  24       3.695222 0.2573080 162.90 3.187133 4.203311
#>  25       3.843346 0.2568431 162.60 3.336168 4.350524
#>  26       3.530254 0.2568467 162.54 3.023068 4.037441
#>  27       3.503787 0.2562348 161.83 2.997792 4.009782
#>  28       3.105117 0.2562914 162.79 2.599033 3.611201
#>  29       3.907397 0.2569825 162.24 3.399935 4.414859
#>  30       3.546053 0.2559495 161.13 3.040605 4.051501
#>  31       4.117931 0.2569224 162.53 3.610594 4.625267
#>  32       4.287392 0.2580407 162.63 3.777850 4.796934
#>  33       3.718718 0.2573236 162.12 3.210580 4.226856
#>  34       3.885943 0.2591319 164.03 3.374279 4.397607
#>  35       4.092327 0.2570324 162.25 3.584767 4.599887
#>  36       3.854670 0.2572876 163.03 3.346625 4.362716
#>  37       3.430916 0.2571593 162.29 2.923106 3.938725
#>  38       3.067100 0.2573586 162.29 2.558897 3.575303
#>  39       3.511730 0.2574858 163.18 3.003296 4.020163
#>  40       3.737128 0.2586285 163.49 3.226445 4.247811
#>  41       3.757774 0.2580288 163.35 3.248272 4.267275
#>  42       3.879338 0.2568476 162.11 3.372139 4.386536
#>  45       4.213687 0.2561950 162.49 3.707786 4.719587
#>  46       3.480188 0.2571277 162.59 2.972447 3.987928
#>  47       4.463420 0.2561192 161.86 3.957654 4.969186
#>  48       3.786229 0.2567359 161.96 3.279248 4.293210
#>  50       2.535141 0.2573069 162.91 2.027054 3.043227
#>  51       3.462136 0.2574433 162.77 2.953777 3.970495
#>  52       3.347574 0.2574401 162.64 2.839218 3.855930
#>  53       3.592387 0.2558585 161.77 3.087134 4.097640
#>  54       3.533987 0.2568994 162.80 3.026702 4.041271
#>  55       2.738430 0.2585452 163.82 2.227920 3.248941
#>  56       2.761946 0.2604520 164.34 2.247682 3.276209
#>  57       3.978624 0.2552671 161.93 3.474542 4.482705
#>  58       2.991384 0.2569200 162.35 2.484048 3.498720
#>  59       3.404556 0.2557690 161.71 2.899478 3.909634
#>  60       3.009904 0.2567573 162.84 2.502901 3.516907
#>  61       3.190840 0.2577702 163.78 2.681858 3.699821
#>  62       3.256187 0.2563370 162.52 2.750007 3.762368
#>  63       3.564469 0.2590750 164.06 3.052917 4.076020
#> 
#> Results are averaged over the levels of: Replication, Column 
#> Degrees-of-freedom method: kenward-roger 
#> Confidence level used: 0.95 

  RCB1<- aov(Dbh~Province+Replication,data = DataExam8.1)
  RCB<- emmeans::emmeans(RCB1,specs="Province")
  Mixed<- emmeans::emmeans(fm8.8,specs ="Province")
  table8.9<-data.frame(as.data.frame(summary(RCB))[c('emmean', 'SE')],
  as.data.frame(summary(Mixed))[c('emmean', 'SE')])
  colnames(table8.9)<- c("RCB(Mean)","RCB(SE)", "Mixed(Mean)", "Mixed(SE)")
  print(table8.9)

#>    RCB(Mean)   RCB(SE) Mixed(Mean) Mixed(SE)
#> 1     1.8500 0.3821484    2.443605 0.2572392
#> 2     3.2400 0.3821484    3.153833 0.2590290
#> 3     2.9375 0.3821484    2.872521 0.2561955
#> 4     2.4975 0.3821484    2.265480 0.2583426
#> 5     3.3425 0.3821484    3.742912 0.2568539
#> 6     3.3750 0.3821484    3.538508 0.2599454
#> 7     3.9825 0.3821484    3.934518 0.2578819
#> 8     2.6350 0.3821484    2.716899 0.2566741
#> 9     2.4650 0.3821484    2.675941 0.2569938
#> 10    2.3950 0.3821484    2.038708 0.2562404
#> 11    2.3875 0.3821484    2.770811 0.2571088
#> 12    2.5775 0.3821484    2.962280 0.2575629
#> 13    2.3425 0.3821484    2.657488 0.2561172
#> 14    3.7650 0.3821484    3.448912 0.2558282
#> 15    3.6000 0.3821484    3.867612 0.2557240
#> 16    3.4450 0.3821484    3.444553 0.2566187
#> 17    3.2550 0.3821484    3.360986 0.2571419
#> 18    3.8300 0.3821484    3.659551 0.2568099
#> 19    3.5525 0.3821484    3.368791 0.2567583
#> 20    3.3225 0.3821484    3.235259 0.2554642
#> 21    3.5475 0.3821484    3.328514 0.2573288
#> 22    3.1675 0.3821484    3.251721 0.2576640
#> 23    3.6425 0.3821484    3.695222 0.2573080
#> 24    4.0050 0.3821484    3.843346 0.2568431
#> 25    3.7350 0.3821484    3.530254 0.2568467
#> 26    3.5175 0.3821484    3.503787 0.2562348
#> 27    2.9925 0.3821484    3.105117 0.2562914
#> 28    4.3350 0.3821484    3.907397 0.2569825
#> 29    3.8225 0.3821484    3.546053 0.2559495
#> 30    4.3875 0.3821484    4.117931 0.2569224
#> 31    4.0425 0.3821484    4.287392 0.2580407
#> 32    3.8050 0.3821484    3.718718 0.2573236
#> 33    3.5850 0.3821484    3.885943 0.2591319
#> 34    4.2625 0.3821484    4.092327 0.2570324
#> 35    4.0025 0.3821484    3.854670 0.2572876
#> 36    3.6500 0.3821484    3.430916 0.2571593
#> 37    3.4600 0.3821484    3.067100 0.2573586
#> 38    3.3575 0.3821484    3.511730 0.2574858
#> 39    4.0175 0.3821484    3.737128 0.2586285
#> 40    3.4900 0.3821484    3.757774 0.2580288
#> 41    3.9075 0.3821484    3.879338 0.2568476
#> 42    3.9550 0.3821484    4.213687 0.2561950
#> 43    3.1975 0.3821484    3.480188 0.2571277
#> 44    4.1700 0.3821484    4.463420 0.2561192
#> 45    4.0425 0.3821484    3.786229 0.2567359
#> 46    2.7625 0.3821484    2.535141 0.2573069
#> 47    3.5575 0.3821484    3.462136 0.2574433
#> 48    3.1075 0.3821484    3.347574 0.2574401
#> 49    3.6275 0.3821484    3.592387 0.2558585
#> 50    3.4775 0.3821484    3.533987 0.2568994
#> 51    2.6125 0.3821484    2.738430 0.2585452
#> 52    2.8700 0.3821484    2.761946 0.2604520
#> 53    4.0350 0.3821484    3.978624 0.2552671
#> 54    3.1225 0.3821484    2.991384 0.2569200
#> 55    3.3450 0.3821484    3.404556 0.2557690
#> 56    3.3675 0.3821484    3.009904 0.2567573
#> 57    2.9925 0.3821484    3.190840 0.2577702
#> 58    3.3450 0.3821484    3.256187 0.2563370
#> 59    3.7000 0.3821484    3.564469 0.2590750