Example 8.1.1 from Experimental Design & Analysis for Tree Improvement

Exam8.1.1 presents the Mixed Effects Analysis of Diameter at breast height (Dbh) of 60 SeedLots under layout of row column design with 6 rows and 10 columns in 18 countries and 59 provinces of 18 selected countries given in Example 8.1.

References

Williams, E.R., Matheson, A.C. and Harwood, C.E. (2002).Experimental Design and Analysis for Tree Improvement. CSIRO Publishing.

See also

Examples

data(DataExam8.1)
library(tidyverse)
library(emmeans)
library(lme4)
fm8.8<- lmer(formula=
     Dbh~1+Replication+Column+Province+(1|Replication:Row)+(1|Replication:Column)
    ,data = DataExam8.1
    ,REML = TRUE
    ,control = lmerControl()
    ,start = NULL
    , verbose = 0L
   #, subset
   #, weights
   #, na.action
   #,offset
    , contrasts = NULL
    , devFunOnly = FALSE)
  anova(fm8.8)

#> Analysis of Variance Table
#>             Df Sum Sq Mean Sq F value
#> Replication  3  2.604 0.86799  4.4054
#> Column       9 28.199 3.13325 15.9027
#> Province    58 55.316 0.95372  4.8406

  summary(fm8.8)

#> Linear mixed model fit by REML ['lmerMod']
#> Formula: Dbh ~ 1 + Replication + Column + Province + (1 | Replication:Row) +  
#>     (1 | Replication:Column)
#>    Data: DataExam8.1
#> 
#> REML criterion at convergence: 358.6
#> 
#> Scaled residuals: 
#>      Min       1Q   Median       3Q      Max 
#> -2.69863 -0.50295  0.04918  0.54501  1.67530 
#> 
#> Random effects:
#>  Groups             Name        Variance Std.Dev.
#>  Replication:Column (Intercept) 0.04607  0.2146  
#>  Replication:Row    (Intercept) 0.05987  0.2447  
#>  Residual                       0.19703  0.4439  
#> Number of obs: 236, groups:  Replication:Column, 40; Replication:Row, 24
#> 
#> Fixed effects:
#>               Estimate Std. Error t value
#> (Intercept)   2.901640   0.305602   9.495
#> Replication2 -0.625047   0.189413  -3.300
#> Replication3 -0.544633   0.189411  -2.875
#> Replication4 -0.252900   0.189447  -1.335
#> Column2       0.003435   0.211350   0.016
#> Column3       0.355786   0.207659   1.713
#> Column4       0.326485   0.208629   1.565
#> Column5       0.094549   0.207378   0.456
#> Column6       0.285696   0.210291   1.359
#> Column7      -0.200472   0.206776  -0.970
#> Column8       0.009088   0.208545   0.044
#> Column9      -0.325797   0.210140  -1.550
#> Column10     -1.572672   0.209442  -7.509
#> Province2     0.710229   0.354793   2.002
#> Province3     0.428916   0.345468   1.242
#> Province4    -0.178124   0.354409  -0.503
#> Province5     1.299307   0.338097   3.843
#> Province6     1.094903   0.355427   3.081
#> Province7     1.490913   0.347099   4.295
#> Province8     0.273295   0.349186   0.783
#> Province10    0.232337   0.351343   0.661
#> Province11   -0.404897   0.347119  -1.166
#> Province12    0.327206   0.343284   0.953
#> Province13    0.518675   0.354027   1.465
#> Province14    0.213883   0.348989   0.613
#> Province15    1.005307   0.347914   2.890
#> Province16    1.424008   0.348833   4.082
#> Province17    1.000948   0.348453   2.873
#> Province18    0.917381   0.353613   2.594
#> Province19    1.215946   0.341866   3.557
#> Province20    0.925186   0.354913   2.607
#> Province21    0.791654   0.347596   2.278
#> Province22    0.884909   0.353644   2.502
#> Province23    0.808116   0.349819   2.310
#> Province24    1.251617   0.353276   3.543
#> Province25    1.399741   0.355797   3.934
#> Province26    1.086650   0.353742   3.072
#> Province27    1.060182   0.353456   2.999
#> Province28    0.661512   0.339238   1.950
#> Province29    1.463792   0.346371   4.226
#> Province30    1.102448   0.354548   3.109
#> Province31    1.674326   0.354615   4.722
#> Province32    1.843787   0.348492   5.291
#> Province33    1.275113   0.356381   3.578
#> Province34    1.442338   0.343664   4.197
#> Province35    1.648722   0.352634   4.675
#> Province36    1.411066   0.348875   4.045
#> Province37    0.987311   0.349879   2.822
#> Province38    0.623495   0.343593   1.815
#> Province39    1.068125   0.348690   3.063
#> Province40    1.293523   0.349723   3.699
#> Province41    1.314169   0.355238   3.699
#> Province42    1.435733   0.349513   4.108
#> Province45    1.770082   0.346258   5.112
#> Province46    1.036583   0.344447   3.009
#> Province47    2.019815   0.348256   5.800
#> Province48    1.342624   0.337249   3.981
#> Province50    0.091536   0.345515   0.265
#> Province51    1.018531   0.346239   2.942
#> Province52    0.903969   0.342882   2.636
#> Province53    1.148782   0.350055   3.282
#> Province54    1.090382   0.354716   3.074
#> Province55    0.294825   0.353085   0.835
#> Province56    0.318341   0.352053   0.904
#> Province57    1.535019   0.342900   4.477
#> Province58    0.547779   0.353748   1.549
#> Province59    0.960952   0.343013   2.801
#> Province60    0.566299   0.350254   1.617
#> Province61    0.747235   0.353647   2.113
#> Province62    0.812583   0.348005   2.335
#> Province63    1.120864   0.357348   3.137

#> 
#> Correlation matrix not shown by default, as p = 71 > 12.
#> Use print(x, correlation=TRUE)  or
#> 	 vcov(x)	 if you need it

  anova(fm8.8, ddf = "Kenward-Roger")

#> Analysis of Variance Table
#>             Df Sum Sq Mean Sq F value
#> Replication  3  2.604 0.86799  4.4054
#> Column       9 28.199 3.13325 15.9027
#> Province    58 55.316 0.95372  4.8406

  emmeans::emmeans(fm8.8,specs ="Replication")

#>  Replication   emmean        SE    df lower.CL upper.CL
#>  1           3.754334 0.1339419 22.45 3.476876 4.031791
#>  2           3.129287 0.1339432 22.45 2.851827 3.406746
#>  3           3.209701 0.1339369 22.45 2.932253 3.487148
#>  4           3.501433 0.1339879 22.47 3.223893 3.778974
#> 
#> Results are averaged over the levels of: Column, Province 
#> Degrees-of-freedom method: kenward-roger 
#> Confidence level used: 0.95 

  emmeans::emmeans(fm8.8,specs ="Column")

#>  Column   emmean        SE    df lower.CL upper.CL
#>  1      3.501079 0.1550087 26.78 3.182905 3.819253
#>  2      3.504513 0.1564558 27.80 3.183926 3.825101
#>  3      3.856865 0.1548582 26.74 3.538979 4.174751
#>  4      3.827564 0.1549569 26.79 3.509504 4.145624
#>  5      3.595628 0.1549173 26.76 3.277633 3.913622
#>  6      3.786775 0.1573353 28.23 3.464606 4.108943
#>  7      3.300607 0.1548099 26.71 2.982802 3.618411
#>  8      3.510167 0.1569450 28.04 3.188699 3.831635
#>  9      3.175282 0.1548719 26.75 2.857371 3.493194
#>  10     1.928407 0.1572168 28.25 1.606492 2.250321
#> 
#> Results are averaged over the levels of: Replication, Province 
#> Degrees-of-freedom method: kenward-roger 
#> Confidence level used: 0.95 

  emmeans::emmeans(fm8.8,specs ="Province")

#>  Province   emmean        SE     df lower.CL upper.CL
#>  1        2.443605 0.2572392 162.87 1.935651 2.951559
#>  2        3.153833 0.2590290 163.59 2.642362 3.665305
#>  3        2.872521 0.2561955 162.14 2.366611 3.378431
#>  4        2.265480 0.2583426 163.44 1.755361 2.775600
#>  5        3.742912 0.2568539 162.34 3.235707 4.250118
#>  6        3.538508 0.2599454 164.54 3.025249 4.051766
#>  7        3.934518 0.2578819 163.38 3.425307 4.443729
#>  8        2.716899 0.2566741 162.67 2.210057 3.223742
#>  10       2.675941 0.2569938 162.56 2.168465 3.183418
#>  11       2.038708 0.2562404 162.80 1.532724 2.544691
#>  12       2.770811 0.2571088 162.50 2.263106 3.278516
#>  13       2.962280 0.2575629 163.21 2.453695 3.470865
#>  14       2.657488 0.2561172 161.82 2.151725 3.163250
#>  15       3.448912 0.2558282 161.61 2.943715 3.954109
#>  16       3.867612 0.2557240 162.31 3.362637 4.372587
#>  17       3.444553 0.2566187 162.56 2.937817 3.951289
#>  18       3.360986 0.2571419 162.85 2.853224 3.868748
#>  19       3.659551 0.2568099 162.48 3.152436 4.166667
#>  20       3.368791 0.2567583 162.54 2.861779 3.875803
#>  21       3.235259 0.2554642 161.38 2.730775 3.739743
#>  22       3.328514 0.2573288 163.16 2.820390 3.836638
#>  23       3.251721 0.2576640 163.33 2.742939 3.760503
#>  24       3.695222 0.2573080 162.90 3.187133 4.203311
#>  25       3.843346 0.2568431 162.60 3.336168 4.350524
#>  26       3.530254 0.2568467 162.54 3.023068 4.037441
#>  27       3.503787 0.2562348 161.83 2.997792 4.009782
#>  28       3.105117 0.2562914 162.79 2.599033 3.611201
#>  29       3.907397 0.2569825 162.24 3.399935 4.414859
#>  30       3.546053 0.2559495 161.13 3.040605 4.051501
#>  31       4.117931 0.2569224 162.53 3.610594 4.625267
#>  32       4.287392 0.2580407 162.63 3.777850 4.796934
#>  33       3.718718 0.2573236 162.12 3.210580 4.226856
#>  34       3.885943 0.2591319 164.03 3.374279 4.397607
#>  35       4.092327 0.2570324 162.25 3.584767 4.599887
#>  36       3.854670 0.2572876 163.03 3.346625 4.362716
#>  37       3.430916 0.2571593 162.29 2.923106 3.938725
#>  38       3.067100 0.2573586 162.29 2.558897 3.575303
#>  39       3.511730 0.2574858 163.18 3.003296 4.020163
#>  40       3.737128 0.2586285 163.49 3.226445 4.247811
#>  41       3.757774 0.2580288 163.35 3.248272 4.267275
#>  42       3.879338 0.2568476 162.11 3.372139 4.386536
#>  45       4.213687 0.2561950 162.49 3.707786 4.719587
#>  46       3.480188 0.2571277 162.59 2.972447 3.987928
#>  47       4.463420 0.2561192 161.86 3.957654 4.969186
#>  48       3.786229 0.2567359 161.96 3.279248 4.293210
#>  50       2.535141 0.2573069 162.91 2.027054 3.043227
#>  51       3.462136 0.2574433 162.77 2.953777 3.970495
#>  52       3.347574 0.2574401 162.64 2.839218 3.855930
#>  53       3.592387 0.2558585 161.77 3.087134 4.097640
#>  54       3.533987 0.2568994 162.80 3.026702 4.041271
#>  55       2.738430 0.2585452 163.82 2.227920 3.248941
#>  56       2.761946 0.2604520 164.34 2.247682 3.276209
#>  57       3.978624 0.2552671 161.93 3.474542 4.482705
#>  58       2.991384 0.2569200 162.35 2.484048 3.498720
#>  59       3.404556 0.2557690 161.71 2.899478 3.909634
#>  60       3.009904 0.2567573 162.84 2.502901 3.516907
#>  61       3.190840 0.2577702 163.78 2.681858 3.699821
#>  62       3.256187 0.2563370 162.52 2.750007 3.762368
#>  63       3.564469 0.2590750 164.06 3.052917 4.076020
#> 
#> Results are averaged over the levels of: Replication, Column 
#> Degrees-of-freedom method: kenward-roger 
#> Confidence level used: 0.95 

  RCB1<- aov(Dbh~Province+Replication,data = DataExam8.1)
  RCB<- emmeans::emmeans(RCB1,specs="Province")
  Mixed<- emmeans::emmeans(fm8.8,specs ="Province")
  table8.9<-data.frame(as.data.frame(summary(RCB))[c('emmean', 'SE')],
  as.data.frame(summary(Mixed))[c('emmean', 'SE')])
  colnames(table8.9)<- c("RCB(Mean)","RCB(SE)", "Mixed(Mean)", "Mixed(SE)")
  print(table8.9)

#>    RCB(Mean)   RCB(SE) Mixed(Mean) Mixed(SE)
#> 1     1.8500 0.3821484    2.443605 0.2572392
#> 2     3.2400 0.3821484    3.153833 0.2590290
#> 3     2.9375 0.3821484    2.872521 0.2561955
#> 4     2.4975 0.3821484    2.265480 0.2583426
#> 5     3.3425 0.3821484    3.742912 0.2568539
#> 6     3.3750 0.3821484    3.538508 0.2599454
#> 7     3.9825 0.3821484    3.934518 0.2578819
#> 8     2.6350 0.3821484    2.716899 0.2566741
#> 9     2.4650 0.3821484    2.675941 0.2569938
#> 10    2.3950 0.3821484    2.038708 0.2562404
#> 11    2.3875 0.3821484    2.770811 0.2571088
#> 12    2.5775 0.3821484    2.962280 0.2575629
#> 13    2.3425 0.3821484    2.657488 0.2561172
#> 14    3.7650 0.3821484    3.448912 0.2558282
#> 15    3.6000 0.3821484    3.867612 0.2557240
#> 16    3.4450 0.3821484    3.444553 0.2566187
#> 17    3.2550 0.3821484    3.360986 0.2571419
#> 18    3.8300 0.3821484    3.659551 0.2568099
#> 19    3.5525 0.3821484    3.368791 0.2567583
#> 20    3.3225 0.3821484    3.235259 0.2554642
#> 21    3.5475 0.3821484    3.328514 0.2573288
#> 22    3.1675 0.3821484    3.251721 0.2576640
#> 23    3.6425 0.3821484    3.695222 0.2573080
#> 24    4.0050 0.3821484    3.843346 0.2568431
#> 25    3.7350 0.3821484    3.530254 0.2568467
#> 26    3.5175 0.3821484    3.503787 0.2562348
#> 27    2.9925 0.3821484    3.105117 0.2562914
#> 28    4.3350 0.3821484    3.907397 0.2569825
#> 29    3.8225 0.3821484    3.546053 0.2559495
#> 30    4.3875 0.3821484    4.117931 0.2569224
#> 31    4.0425 0.3821484    4.287392 0.2580407
#> 32    3.8050 0.3821484    3.718718 0.2573236
#> 33    3.5850 0.3821484    3.885943 0.2591319
#> 34    4.2625 0.3821484    4.092327 0.2570324
#> 35    4.0025 0.3821484    3.854670 0.2572876
#> 36    3.6500 0.3821484    3.430916 0.2571593
#> 37    3.4600 0.3821484    3.067100 0.2573586
#> 38    3.3575 0.3821484    3.511730 0.2574858
#> 39    4.0175 0.3821484    3.737128 0.2586285
#> 40    3.4900 0.3821484    3.757774 0.2580288
#> 41    3.9075 0.3821484    3.879338 0.2568476
#> 42    3.9550 0.3821484    4.213687 0.2561950
#> 43    3.1975 0.3821484    3.480188 0.2571277
#> 44    4.1700 0.3821484    4.463420 0.2561192
#> 45    4.0425 0.3821484    3.786229 0.2567359
#> 46    2.7625 0.3821484    2.535141 0.2573069
#> 47    3.5575 0.3821484    3.462136 0.2574433
#> 48    3.1075 0.3821484    3.347574 0.2574401
#> 49    3.6275 0.3821484    3.592387 0.2558585
#> 50    3.4775 0.3821484    3.533987 0.2568994
#> 51    2.6125 0.3821484    2.738430 0.2585452
#> 52    2.8700 0.3821484    2.761946 0.2604520
#> 53    4.0350 0.3821484    3.978624 0.2552671
#> 54    3.1225 0.3821484    2.991384 0.2569200
#> 55    3.3450 0.3821484    3.404556 0.2557690
#> 56    3.3675 0.3821484    3.009904 0.2567573
#> 57    2.9925 0.3821484    3.190840 0.2577702
#> 58    3.3450 0.3821484    3.256187 0.2563370
#> 59    3.7000 0.3821484    3.564469 0.2590750