Example 8.1.1 from Experimental Design & Analysis for Tree Improvement

Exam8.1.1 presents the Mixed Effects Analysis of Diameter at breast height (Dbh) of 60 SeedLots under layout of row column design with 6 rows and 10 columns in 18 countries and 59 provinces of 18 selected countries given in Example 8.1.

References

Williams, E.R., Matheson, A.C. and Harwood, C.E. (2023).Experimental Design and Analysis for Tree Improvement. CSIRO Publishing.

See also

Author

Muhammad Yaseen (myaseen208@gmail.com)

Sami Ullah (samiullahuos@gmail.com)

Examples

data(DataExam8.1)
library(tidyverse)
library(emmeans)
library(lmerTest)
#> 
#> Attaching package: 'lmerTest'
#> The following object is masked from 'package:lme4':
#> 
#>     lmer
#> The following object is masked from 'package:stats':
#> 
#>     step
library(lme4)

# Pg. 155
fm8.8 <-
    lmerTest::lmer(
            formula    = Dbh ~ 1 + Replication + Column + Province + (1|Replication:Row) + (1|Replication:Column)
          , data       = DataExam8.1
          , REML       = TRUE
          , control    = lmerControl()
          , start      = NULL
          , verbose    = 0L
        # , subset
        # , weights
        # , na.action
        # , offset
          , contrasts  = NULL
          , devFunOnly = FALSE
          )

 # Pg. 157

  summary(fm8.8)
#> Linear mixed model fit by REML. t-tests use Satterthwaite's method [
#> lmerModLmerTest]
#> Formula: Dbh ~ 1 + Replication + Column + Province + (1 | Replication:Row) +  
#>     (1 | Replication:Column)
#>    Data: DataExam8.1
#> 
#> REML criterion at convergence: 358.6
#> 
#> Scaled residuals: 
#>      Min       1Q   Median       3Q      Max 
#> -2.69863 -0.50295  0.04918  0.54501  1.67530 
#> 
#> Random effects:
#>  Groups             Name        Variance Std.Dev.
#>  Replication:Column (Intercept) 0.04607  0.2146  
#>  Replication:Row    (Intercept) 0.05987  0.2447  
#>  Residual                       0.19703  0.4439  
#> Number of obs: 236, groups:  Replication:Column, 40; Replication:Row, 24
#> 
#> Fixed effects:
#>                Estimate Std. Error         df t value Pr(>|t|)    
#> (Intercept)    2.901640   0.305602 112.174435   9.495 4.78e-16 ***
#> Replication2  -0.625047   0.189413  21.418195  -3.300 0.003347 ** 
#> Replication3  -0.544633   0.189411  21.416868  -2.875 0.008940 ** 
#> Replication4  -0.252900   0.189447  21.426469  -1.335 0.195907    
#> Column2        0.003435   0.211350  24.743600   0.016 0.987165    
#> Column3        0.355786   0.207659  23.070583   1.713 0.100061    
#> Column4        0.326485   0.208629  23.495309   1.565 0.130975    
#> Column5        0.094549   0.207378  22.903081   0.456 0.652737    
#> Column6        0.285696   0.210291  24.193582   1.359 0.186815    
#> Column7       -0.200472   0.206776  22.692169  -0.970 0.342509    
#> Column8        0.009088   0.208545  23.415421   0.044 0.965610    
#> Column9       -0.325797   0.210140  24.192520  -1.550 0.134035    
#> Column10      -1.572672   0.209443  23.927072  -7.509 9.69e-08 ***
#> Province2      0.710228   0.354793 155.494120   2.002 0.047044 *  
#> Province3      0.428916   0.345468 144.578828   1.242 0.216413    
#> Province4     -0.178125   0.354409 156.300767  -0.503 0.615956    
#> Province5      1.299307   0.338097 141.585860   3.843 0.000183 ***
#> Province6      1.094903   0.355427 157.842574   3.081 0.002438 ** 
#> Province7      1.490913   0.347099 149.238144   4.295 3.13e-05 ***
#> Province8      0.273295   0.349186 147.793242   0.783 0.435078    
#> Province10     0.232336   0.351343 151.791245   0.661 0.509435    
#> Province11    -0.404897   0.347119 147.856819  -1.166 0.245310    
#> Province12     0.327206   0.343284 145.038305   0.953 0.342093    
#> Province13     0.518675   0.354027 155.108827   1.465 0.144925    
#> Province14     0.213883   0.348989 149.259583   0.613 0.540900    
#> Province15     1.005307   0.347914 148.057611   2.890 0.004438 ** 
#> Province16     1.424007   0.348833 151.271059   4.082 7.20e-05 ***
#> Province17     1.000948   0.348453 150.213176   2.873 0.004661 ** 
#> Province18     0.917381   0.353613 152.316488   2.594 0.010404 *  
#> Province19     1.215946   0.341866 143.111466   3.557 0.000510 ***
#> Province20     0.925186   0.354913 156.066473   2.607 0.010025 *  
#> Province21     0.791654   0.347596 146.260213   2.278 0.024205 *  
#> Province22     0.884909   0.353644 154.930362   2.502 0.013378 *  
#> Province23     0.808116   0.349819 149.153558   2.310 0.022254 *  
#> Province24     1.251617   0.353276 154.758967   3.543 0.000523 ***
#> Province25     1.399741   0.355797 152.887956   3.934 0.000126 ***
#> Province26     1.086649   0.353742 154.340664   3.072 0.002515 ** 
#> Province27     1.060182   0.353456 150.189885   2.999 0.003167 ** 
#> Province28     0.661512   0.339238 143.290317   1.950 0.053129 .  
#> Province29     1.463791   0.346371 149.511576   4.226 4.12e-05 ***
#> Province30     1.102448   0.354548 151.824553   3.109 0.002239 ** 
#> Province31     1.674326   0.354615 155.163742   4.722 5.20e-06 ***
#> Province32     1.843787   0.348492 148.355084   5.291 4.29e-07 ***
#> Province33     1.275113   0.356381 151.939990   3.578 0.000465 ***
#> Province34     1.442338   0.343664 149.425923   4.197 4.62e-05 ***
#> Province35     1.648722   0.352634 150.435418   4.675 6.47e-06 ***
#> Province36     1.411065   0.348875 150.116792   4.045 8.35e-05 ***
#> Province37     0.987311   0.349879 148.603812   2.822 0.005428 ** 
#> Province38     0.623495   0.343593 144.184979   1.815 0.071659 .  
#> Province39     1.068125   0.348690 151.447236   3.063 0.002592 ** 
#> Province40     1.293523   0.349723 151.955074   3.699 0.000302 ***
#> Province41     1.314169   0.355238 156.069514   3.699 0.000299 ***
#> Province42     1.435733   0.349513 147.587432   4.108 6.60e-05 ***
#> Province45     1.770082   0.346258 149.917649   5.112 9.58e-07 ***
#> Province46     1.036583   0.344447 143.357116   3.009 0.003093 ** 
#> Province47     2.019815   0.348256 150.703961   5.800 3.77e-08 ***
#> Province48     1.342624   0.337249 139.823992   3.981 0.000110 ***
#> Province50     0.091536   0.345515 146.096465   0.265 0.791440    
#> Province51     1.018531   0.346239 144.865161   2.942 0.003802 ** 
#> Province52     0.903969   0.342882 145.406584   2.636 0.009289 ** 
#> Province53     1.148782   0.350055 150.225870   3.282 0.001283 ** 
#> Province54     1.090382   0.354716 153.881408   3.074 0.002500 ** 
#> Province55     0.294825   0.353085 156.191083   0.835 0.404994    
#> Province56     0.318341   0.352053 152.217429   0.904 0.367296    
#> Province57     1.535019   0.342899 141.444724   4.477 1.55e-05 ***
#> Province58     0.547779   0.353748 154.315319   1.549 0.123550    
#> Province59     0.960952   0.343013 145.475049   2.801 0.005779 ** 
#> Province60     0.566299   0.350254 148.687871   1.617 0.108036    
#> Province61     0.747235   0.353647 157.166285   2.113 0.036186 *  
#> Province62     0.812583   0.348005 150.316512   2.335 0.020868 *  
#> Province63     1.120864   0.357348 154.538269   3.137 0.002047 ** 
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Correlation matrix not shown by default, as p = 71 > 12.
#> Use print(x, correlation=TRUE)  or
#>     vcov(x)        if you need it
  summary(fm8.8)$varcor
#>  Groups             Name        Std.Dev.
#>  Replication:Column (Intercept) 0.21464 
#>  Replication:Row    (Intercept) 0.24469 
#>  Residual                       0.44388 

  anova(fm8.8)
#> Type III Analysis of Variance Table with Satterthwaite's method
#>             Sum Sq Mean Sq NumDF   DenDF F value    Pr(>F)    
#> Replication  2.695 0.89819     3  21.422  4.5587   0.01281 *  
#> Column      24.882 2.76470     9  23.702 14.0321 1.652e-07 ***
#> Province    55.316 0.95372    58 137.058  4.8406 1.916e-14 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
  anova(fm8.8, ddf = "Kenward-Roger")
#> Type III Analysis of Variance Table with Kenward-Roger's method
#>             Sum Sq Mean Sq NumDF   DenDF F value    Pr(>F)    
#> Replication  2.694 0.89811     3  22.453  4.5583   0.01226 *  
#> Column      24.808 2.75650     9  22.986 13.9898 2.268e-07 ***
#> Province    54.691 0.94295    58 134.064  4.7842 4.195e-14 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

  library(car)
#> Loading required package: carData
#> 
#> Attaching package: 'car'
#> The following object is masked from 'package:dplyr':
#> 
#>     recode
#> The following object is masked from 'package:purrr':
#> 
#>     some
  Anova(fm8.8, type = "II", test.statistic = "Chisq")
#> Analysis of Deviance Table (Type II Wald chisquare tests)
#> 
#> Response: Dbh
#>               Chisq Df Pr(>Chisq)    
#> Replication  13.676  3   0.003381 ** 
#> Column      126.289  9  < 2.2e-16 ***
#> Province    280.754 58  < 2.2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

  emmeans(fm8.8, specs = "Replication")
#>  Replication emmean    SE   df lower.CL upper.CL
#>  1             3.75 0.134 22.4     3.48     4.03
#>  2             3.13 0.134 22.4     2.85     3.41
#>  3             3.21 0.134 22.4     2.93     3.49
#>  4             3.50 0.134 22.5     3.22     3.78
#> 
#> Results are averaged over the levels of: Column, Province 
#> Degrees-of-freedom method: kenward-roger 
#> Confidence level used: 0.95 
  emmeans(fm8.8, specs = "Column")
#>  Column emmean    SE   df lower.CL upper.CL
#>  1        3.50 0.155 26.8     3.18     3.82
#>  2        3.50 0.156 27.8     3.18     3.83
#>  3        3.86 0.155 26.7     3.54     4.17
#>  4        3.83 0.155 26.8     3.51     4.15
#>  5        3.60 0.155 26.8     3.28     3.91
#>  6        3.79 0.157 28.2     3.46     4.11
#>  7        3.30 0.155 26.7     2.98     3.62
#>  8        3.51 0.157 28.0     3.19     3.83
#>  9        3.18 0.155 26.8     2.86     3.49
#>  10       1.93 0.157 28.2     1.61     2.25
#> 
#> Results are averaged over the levels of: Replication, Province 
#> Degrees-of-freedom method: kenward-roger 
#> Confidence level used: 0.95 
  emmeans(fm8.8, specs = "Province")
#>  Province emmean    SE  df lower.CL upper.CL
#>  1          2.44 0.257 163     1.94     2.95
#>  2          3.15 0.259 164     2.64     3.67
#>  3          2.87 0.256 162     2.37     3.38
#>  4          2.27 0.258 163     1.76     2.78
#>  5          3.74 0.257 162     3.24     4.25
#>  6          3.54 0.260 165     3.03     4.05
#>  7          3.93 0.258 163     3.43     4.44
#>  8          2.72 0.257 163     2.21     3.22
#>  10         2.68 0.257 163     2.17     3.18
#>  11         2.04 0.256 163     1.53     2.54
#>  12         2.77 0.257 162     2.26     3.28
#>  13         2.96 0.258 163     2.45     3.47
#>  14         2.66 0.256 162     2.15     3.16
#>  15         3.45 0.256 162     2.94     3.95
#>  16         3.87 0.256 162     3.36     4.37
#>  17         3.44 0.257 163     2.94     3.95
#>  18         3.36 0.257 163     2.85     3.87
#>  19         3.66 0.257 162     3.15     4.17
#>  20         3.37 0.257 163     2.86     3.88
#>  21         3.24 0.255 161     2.73     3.74
#>  22         3.33 0.257 163     2.82     3.84
#>  23         3.25 0.258 163     2.74     3.76
#>  24         3.70 0.257 163     3.19     4.20
#>  25         3.84 0.257 163     3.34     4.35
#>  26         3.53 0.257 163     3.02     4.04
#>  27         3.50 0.256 162     3.00     4.01
#>  28         3.11 0.256 163     2.60     3.61
#>  29         3.91 0.257 162     3.40     4.41
#>  30         3.55 0.256 161     3.04     4.05
#>  31         4.12 0.257 163     3.61     4.63
#>  32         4.29 0.258 163     3.78     4.80
#>  33         3.72 0.257 162     3.21     4.23
#>  34         3.89 0.259 164     3.37     4.40
#>  35         4.09 0.257 162     3.58     4.60
#>  36         3.85 0.257 163     3.35     4.36
#>  37         3.43 0.257 162     2.92     3.94
#>  38         3.07 0.257 162     2.56     3.58
#>  39         3.51 0.257 163     3.00     4.02
#>  40         3.74 0.259 163     3.23     4.25
#>  41         3.76 0.258 163     3.25     4.27
#>  42         3.88 0.257 162     3.37     4.39
#>  45         4.21 0.256 162     3.71     4.72
#>  46         3.48 0.257 163     2.97     3.99
#>  47         4.46 0.256 162     3.96     4.97
#>  48         3.79 0.257 162     3.28     4.29
#>  50         2.54 0.257 163     2.03     3.04
#>  51         3.46 0.257 163     2.95     3.97
#>  52         3.35 0.257 163     2.84     3.86
#>  53         3.59 0.256 162     3.09     4.10
#>  54         3.53 0.257 163     3.03     4.04
#>  55         2.74 0.259 164     2.23     3.25
#>  56         2.76 0.260 164     2.25     3.28
#>  57         3.98 0.255 162     3.47     4.48
#>  58         2.99 0.257 162     2.48     3.50
#>  59         3.40 0.256 162     2.90     3.91
#>  60         3.01 0.257 163     2.50     3.52
#>  61         3.19 0.258 164     2.68     3.70
#>  62         3.26 0.256 163     2.75     3.76
#>  63         3.56 0.259 164     3.05     4.08
#> 
#> Results are averaged over the levels of: Replication, Column 
#> Degrees-of-freedom method: kenward-roger 
#> Confidence level used: 0.95 

 # Pg. 161
  RCB1  <- aov(Dbh~Province+Replication,data = DataExam8.1)
  RCB   <-
          emmeans(RCB1, specs = "Province") %>%
          as_tibble()
  Mixed <-
          emmeans(fm8.8, specs = "Province") %>%
          as_tibble()

  table8.9 <- left_join(RCB, Mixed, by = "Province", suffix = c(".RCBD", ".Mixed"))
  print(table8.9)
#> # A tibble: 59 × 11
#>    Province emmean.RCBD SE.RCBD df.RCBD lower.CL.RCBD upper.CL.RCBD emmean.Mixed
#>    <fct>          <dbl>   <dbl>   <dbl>         <dbl>         <dbl>        <dbl>
#>  1 1               1.85   0.382     174          1.10          2.60         2.44
#>  2 2               3.24   0.382     174          2.49          3.99         3.15
#>  3 3               2.94   0.382     174          2.18          3.69         2.87
#>  4 4               2.50   0.382     174          1.74          3.25         2.27
#>  5 5               3.34   0.382     174          2.59          4.10         3.74
#>  6 6               3.37   0.382     174          2.62          4.13         3.54
#>  7 7               3.98   0.382     174          3.23          4.74         3.93
#>  8 8               2.64   0.382     174          1.88          3.39         2.72
#>  9 10              2.47   0.382     174          1.71          3.22         2.68
#> 10 11              2.40   0.382     174          1.64          3.15         2.04
#> # ℹ 49 more rows
#> # ℹ 4 more variables: SE.Mixed <dbl>, df.Mixed <dbl>, lower.CL.Mixed <dbl>,
#> #   upper.CL.Mixed <dbl>