Example 8.1.1 from Experimental Design & Analysis for Tree Improvement

Exam8.1.1 presents the Mixed Effects Analysis of Diameter at breast height (Dbh) of 60 SeedLots under layout of row column design with 6 rows and 10 columns in 18 countries and 59 provinces of 18 selected countries given in Example 8.1.

References

Williams, E.R., Matheson, A.C. and Harwood, C.E. (2002).Experimental Design and Analysis for Tree Improvement. CSIRO Publishing.

See also

Author

Muhammad Yaseen (myaseen208@gmail.com)

Sami Ullah (samiullahuos@gmail.com)

Examples

data(DataExam8.1)
library(tidyverse)
library(emmeans)
library(lme4)
fm8.8<- lmer(formula=
     Dbh~1+Replication+Column+Province+(1|Replication:Row)+(1|Replication:Column)
    ,data = DataExam8.1
    ,REML = TRUE
    ,control = lmerControl()
    ,start = NULL
    , verbose = 0L
   #, subset
   #, weights
   #, na.action
   #,offset
    , contrasts = NULL
    , devFunOnly = FALSE)
  anova(fm8.8)
#> Analysis of Variance Table
#>             npar Sum Sq Mean Sq F value
#> Replication    3  2.604 0.86799  4.4054
#> Column         9 28.199 3.13324 15.9027
#> Province      58 55.316 0.95372  4.8406
  summary(fm8.8)
#> Linear mixed model fit by REML ['lmerMod']
#> Formula: Dbh ~ 1 + Replication + Column + Province + (1 | Replication:Row) +  
#>     (1 | Replication:Column)
#>    Data: DataExam8.1
#> 
#> REML criterion at convergence: 358.6
#> 
#> Scaled residuals: 
#>      Min       1Q   Median       3Q      Max 
#> -2.69863 -0.50295  0.04918  0.54501  1.67530 
#> 
#> Random effects:
#>  Groups             Name        Variance Std.Dev.
#>  Replication:Column (Intercept) 0.04607  0.2146  
#>  Replication:Row    (Intercept) 0.05987  0.2447  
#>  Residual                       0.19703  0.4439  
#> Number of obs: 236, groups:  Replication:Column, 40; Replication:Row, 24
#> 
#> Fixed effects:
#>               Estimate Std. Error t value
#> (Intercept)   2.901640   0.305602   9.495
#> Replication2 -0.625047   0.189413  -3.300
#> Replication3 -0.544633   0.189411  -2.875
#> Replication4 -0.252900   0.189447  -1.335
#> Column2       0.003435   0.211350   0.016
#> Column3       0.355786   0.207659   1.713
#> Column4       0.326485   0.208629   1.565
#> Column5       0.094549   0.207378   0.456
#> Column6       0.285696   0.210291   1.359
#> Column7      -0.200472   0.206776  -0.970
#> Column8       0.009088   0.208545   0.044
#> Column9      -0.325797   0.210140  -1.550
#> Column10     -1.572672   0.209443  -7.509
#> Province2     0.710228   0.354793   2.002
#> Province3     0.428916   0.345468   1.242
#> Province4    -0.178125   0.354409  -0.503
#> Province5     1.299307   0.338097   3.843
#> Province6     1.094903   0.355427   3.081
#> Province7     1.490913   0.347099   4.295
#> Province8     0.273295   0.349186   0.783
#> Province10    0.232336   0.351343   0.661
#> Province11   -0.404897   0.347119  -1.166
#> Province12    0.327206   0.343284   0.953
#> Province13    0.518675   0.354027   1.465
#> Province14    0.213883   0.348989   0.613
#> Province15    1.005307   0.347914   2.890
#> Province16    1.424007   0.348833   4.082
#> Province17    1.000948   0.348453   2.873
#> Province18    0.917381   0.353613   2.594
#> Province19    1.215946   0.341866   3.557
#> Province20    0.925186   0.354913   2.607
#> Province21    0.791654   0.347596   2.278
#> Province22    0.884909   0.353644   2.502
#> Province23    0.808116   0.349819   2.310
#> Province24    1.251617   0.353276   3.543
#> Province25    1.399741   0.355797   3.934
#> Province26    1.086649   0.353742   3.072
#> Province27    1.060182   0.353456   2.999
#> Province28    0.661512   0.339238   1.950
#> Province29    1.463791   0.346371   4.226
#> Province30    1.102448   0.354548   3.109
#> Province31    1.674326   0.354615   4.722
#> Province32    1.843787   0.348492   5.291
#> Province33    1.275113   0.356381   3.578
#> Province34    1.442338   0.343664   4.197
#> Province35    1.648722   0.352634   4.675
#> Province36    1.411065   0.348875   4.045
#> Province37    0.987311   0.349879   2.822
#> Province38    0.623495   0.343593   1.815
#> Province39    1.068125   0.348690   3.063
#> Province40    1.293523   0.349723   3.699
#> Province41    1.314169   0.355238   3.699
#> Province42    1.435733   0.349513   4.108
#> Province45    1.770082   0.346258   5.112
#> Province46    1.036583   0.344447   3.009
#> Province47    2.019815   0.348256   5.800
#> Province48    1.342624   0.337249   3.981
#> Province50    0.091536   0.345515   0.265
#> Province51    1.018531   0.346239   2.942
#> Province52    0.903969   0.342882   2.636
#> Province53    1.148782   0.350055   3.282
#> Province54    1.090382   0.354716   3.074
#> Province55    0.294825   0.353085   0.835
#> Province56    0.318341   0.352053   0.904
#> Province57    1.535019   0.342899   4.477
#> Province58    0.547779   0.353748   1.549
#> Province59    0.960952   0.343013   2.801
#> Province60    0.566299   0.350254   1.617
#> Province61    0.747235   0.353647   2.113
#> Province62    0.812583   0.348005   2.335
#> Province63    1.120864   0.357348   3.137
#> 
#> Correlation matrix not shown by default, as p = 71 > 12.
#> Use print(x, correlation=TRUE)  or
#>     vcov(x)        if you need it
  anova(fm8.8, ddf = "Kenward-Roger")
#> Warning: additional arguments ignored: 'ddf'
#> Analysis of Variance Table
#>             npar Sum Sq Mean Sq F value
#> Replication    3  2.604 0.86799  4.4054
#> Column         9 28.199 3.13324 15.9027
#> Province      58 55.316 0.95372  4.8406
  emmeans::emmeans(fm8.8,specs ="Replication")
#>  Replication emmean    SE   df lower.CL upper.CL
#>  1             3.75 0.134 22.4     3.48     4.03
#>  2             3.13 0.134 22.4     2.85     3.41
#>  3             3.21 0.134 22.4     2.93     3.49
#>  4             3.50 0.134 22.5     3.22     3.78
#> 
#> Results are averaged over the levels of: Column, Province 
#> Degrees-of-freedom method: kenward-roger 
#> Confidence level used: 0.95 
  emmeans::emmeans(fm8.8,specs ="Column")
#>  Column emmean    SE   df lower.CL upper.CL
#>  1        3.50 0.155 26.8     3.18     3.82
#>  2        3.50 0.156 27.8     3.18     3.83
#>  3        3.86 0.155 26.7     3.54     4.17
#>  4        3.83 0.155 26.8     3.51     4.15
#>  5        3.60 0.155 26.8     3.28     3.91
#>  6        3.79 0.157 28.2     3.46     4.11
#>  7        3.30 0.155 26.7     2.98     3.62
#>  8        3.51 0.157 28.0     3.19     3.83
#>  9        3.18 0.155 26.8     2.86     3.49
#>  10       1.93 0.157 28.2     1.61     2.25
#> 
#> Results are averaged over the levels of: Replication, Province 
#> Degrees-of-freedom method: kenward-roger 
#> Confidence level used: 0.95 
  emmeans::emmeans(fm8.8,specs ="Province")
#>  Province emmean    SE  df lower.CL upper.CL
#>  1          2.44 0.257 163     1.94     2.95
#>  2          3.15 0.259 164     2.64     3.67
#>  3          2.87 0.256 162     2.37     3.38
#>  4          2.27 0.258 163     1.76     2.78
#>  5          3.74 0.257 162     3.24     4.25
#>  6          3.54 0.260 165     3.03     4.05
#>  7          3.93 0.258 163     3.43     4.44
#>  8          2.72 0.257 163     2.21     3.22
#>  10         2.68 0.257 163     2.17     3.18
#>  11         2.04 0.256 163     1.53     2.54
#>  12         2.77 0.257 162     2.26     3.28
#>  13         2.96 0.258 163     2.45     3.47
#>  14         2.66 0.256 162     2.15     3.16
#>  15         3.45 0.256 162     2.94     3.95
#>  16         3.87 0.256 162     3.36     4.37
#>  17         3.44 0.257 163     2.94     3.95
#>  18         3.36 0.257 163     2.85     3.87
#>  19         3.66 0.257 162     3.15     4.17
#>  20         3.37 0.257 163     2.86     3.88
#>  21         3.24 0.255 161     2.73     3.74
#>  22         3.33 0.257 163     2.82     3.84
#>  23         3.25 0.258 163     2.74     3.76
#>  24         3.70 0.257 163     3.19     4.20
#>  25         3.84 0.257 163     3.34     4.35
#>  26         3.53 0.257 163     3.02     4.04
#>  27         3.50 0.256 162     3.00     4.01
#>  28         3.11 0.256 163     2.60     3.61
#>  29         3.91 0.257 162     3.40     4.41
#>  30         3.55 0.256 161     3.04     4.05
#>  31         4.12 0.257 163     3.61     4.63
#>  32         4.29 0.258 163     3.78     4.80
#>  33         3.72 0.257 162     3.21     4.23
#>  34         3.89 0.259 164     3.37     4.40
#>  35         4.09 0.257 162     3.58     4.60
#>  36         3.85 0.257 163     3.35     4.36
#>  37         3.43 0.257 162     2.92     3.94
#>  38         3.07 0.257 162     2.56     3.58
#>  39         3.51 0.257 163     3.00     4.02
#>  40         3.74 0.259 163     3.23     4.25
#>  41         3.76 0.258 163     3.25     4.27
#>  42         3.88 0.257 162     3.37     4.39
#>  45         4.21 0.256 162     3.71     4.72
#>  46         3.48 0.257 163     2.97     3.99
#>  47         4.46 0.256 162     3.96     4.97
#>  48         3.79 0.257 162     3.28     4.29
#>  50         2.54 0.257 163     2.03     3.04
#>  51         3.46 0.257 163     2.95     3.97
#>  52         3.35 0.257 163     2.84     3.86
#>  53         3.59 0.256 162     3.09     4.10
#>  54         3.53 0.257 163     3.03     4.04
#>  55         2.74 0.259 164     2.23     3.25
#>  56         2.76 0.260 164     2.25     3.28
#>  57         3.98 0.255 162     3.47     4.48
#>  58         2.99 0.257 162     2.48     3.50
#>  59         3.40 0.256 162     2.90     3.91
#>  60         3.01 0.257 163     2.50     3.52
#>  61         3.19 0.258 164     2.68     3.70
#>  62         3.26 0.256 163     2.75     3.76
#>  63         3.56 0.259 164     3.05     4.08
#> 
#> Results are averaged over the levels of: Replication, Column 
#> Degrees-of-freedom method: kenward-roger 
#> Confidence level used: 0.95 
  RCB1<- aov(Dbh~Province+Replication,data = DataExam8.1)
  RCB<- emmeans::emmeans(RCB1,specs="Province")
  Mixed<- emmeans::emmeans(fm8.8,specs ="Province")
  table8.9<-data.frame(as.data.frame(summary(RCB))[c('emmean', 'SE')],
  as.data.frame(summary(Mixed))[c('emmean', 'SE')])
  colnames(table8.9)<- c("RCB(Mean)","RCB(SE)", "Mixed(Mean)", "Mixed(SE)")
  print(table8.9)
#>    RCB(Mean)   RCB(SE) Mixed(Mean) Mixed(SE)
#> 1     1.8500 0.3821484    2.443605 0.2572392
#> 2     3.2400 0.3821484    3.153833 0.2590290
#> 3     2.9375 0.3821484    2.872521 0.2561955
#> 4     2.4975 0.3821484    2.265480 0.2583426
#> 5     3.3425 0.3821484    3.742912 0.2568539
#> 6     3.3750 0.3821484    3.538508 0.2599454
#> 7     3.9825 0.3821484    3.934518 0.2578819
#> 8     2.6350 0.3821484    2.716900 0.2566741
#> 9     2.4650 0.3821484    2.675941 0.2569938
#> 10    2.3950 0.3821484    2.038708 0.2562404
#> 11    2.3875 0.3821484    2.770811 0.2571088
#> 12    2.5775 0.3821484    2.962280 0.2575629
#> 13    2.3425 0.3821484    2.657488 0.2561172
#> 14    3.7650 0.3821484    3.448912 0.2558282
#> 15    3.6000 0.3821484    3.867612 0.2557240
#> 16    3.4450 0.3821484    3.444553 0.2566187
#> 17    3.2550 0.3821484    3.360986 0.2571419
#> 18    3.8300 0.3821484    3.659551 0.2568099
#> 19    3.5525 0.3821484    3.368791 0.2567583
#> 20    3.3225 0.3821484    3.235259 0.2554642
#> 21    3.5475 0.3821484    3.328514 0.2573288
#> 22    3.1675 0.3821484    3.251721 0.2576640
#> 23    3.6425 0.3821484    3.695222 0.2573080
#> 24    4.0050 0.3821484    3.843346 0.2568431
#> 25    3.7350 0.3821484    3.530254 0.2568466
#> 26    3.5175 0.3821484    3.503787 0.2562348
#> 27    2.9925 0.3821484    3.105117 0.2562914
#> 28    4.3350 0.3821484    3.907396 0.2569825
#> 29    3.8225 0.3821484    3.546053 0.2559495
#> 30    4.3875 0.3821484    4.117931 0.2569224
#> 31    4.0425 0.3821484    4.287392 0.2580407
#> 32    3.8050 0.3821484    3.718718 0.2573235
#> 33    3.5850 0.3821484    3.885943 0.2591319
#> 34    4.2625 0.3821484    4.092327 0.2570324
#> 35    4.0025 0.3821484    3.854670 0.2572876
#> 36    3.6500 0.3821484    3.430916 0.2571593
#> 37    3.4600 0.3821484    3.067100 0.2573586
#> 38    3.3575 0.3821484    3.511730 0.2574858
#> 39    4.0175 0.3821484    3.737128 0.2586285
#> 40    3.4900 0.3821484    3.757773 0.2580288
#> 41    3.9075 0.3821484    3.879338 0.2568476
#> 42    3.9550 0.3821484    4.213687 0.2561950
#> 43    3.1975 0.3821484    3.480188 0.2571277
#> 44    4.1700 0.3821484    4.463420 0.2561192
#> 45    4.0425 0.3821484    3.786229 0.2567359
#> 46    2.7625 0.3821484    2.535141 0.2573069
#> 47    3.5575 0.3821484    3.462136 0.2574433
#> 48    3.1075 0.3821484    3.347574 0.2574401
#> 49    3.6275 0.3821484    3.592387 0.2558584
#> 50    3.4775 0.3821484    3.533987 0.2568994
#> 51    2.6125 0.3821484    2.738430 0.2585452
#> 52    2.8700 0.3821484    2.761946 0.2604520
#> 53    4.0350 0.3821484    3.978624 0.2552671
#> 54    3.1225 0.3821484    2.991384 0.2569200
#> 55    3.3450 0.3821484    3.404557 0.2557690
#> 56    3.3675 0.3821484    3.009904 0.2567573
#> 57    2.9925 0.3821484    3.190840 0.2577702
#> 58    3.3450 0.3821484    3.256187 0.2563370
#> 59    3.7000 0.3821484    3.564469 0.2590750