Example 8.1.2 from Experimental Design & Analysis for Tree Improvement

Exam8.1.2 presents the Analysis of Nested Seedlot Structure of Diameter at breast height (Dbh) of 60 SeedLots under layout of row column design with 6 rows and 10 columns in 18 countries and 59 provinces of 18 selected countries given in Example 8.1.

References

Williams, E.R., Matheson, A.C. and Harwood, C.E. (2002).Experimental Design and Analysis for Tree Improvement. CSIRO Publishing.

See also

Author

Muhammad Yaseen (myaseen208@gmail.com)

Sami Ullah (samiullahuos@gmail.com)

Examples

data(DataExam8.1)
library(tidyverse)
library(emmeans)
  fm8.11 <- aov(formula      =
  Dbh~Country*Province
  ,data         = DataExam8.1
  #, subset
  #, weights
  #, na.action
  , method      = "qr"
  , model       = TRUE
  , x           = FALSE
  , y           = FALSE
  , qr          = TRUE
  , singular.ok = TRUE
  , contrasts   = NULL
  )
  b<-anova(fm8.11)
     Res                     <- length(b[["Sum Sq"]])
     df                      <- 119
     MSS                     <- 0.1951
     b[["Df"]][Res]          <- df
     b[["Sum Sq"]][Res]      <- MSS*df
     b[["Mean Sq"]][Res]     <- b[["Sum Sq"]][Res]/b[["Df"]][Res]
     b[["F value"]][1:Res-1] <- b[["Mean Sq"]][1:Res-1]/b[["Mean Sq"]][Res]
     b[["Pr(>F)"]][Res-1]     <- df(b[["F value"]][Res-1],b[["Df"]][Res-1],b[["Df"]][Res])
     b
#> Analysis of Variance Table
#> 
#> Response: Dbh
#>            Df Sum Sq Mean Sq F value    Pr(>F)    
#> Country    17 54.651  3.2148 16.4777 1.177e-08 ***
#> Province   41 18.653  0.4549  2.3318  0.001442 ** 
#> Residuals 119 23.217  0.1951                      
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
  library(emmeans)
  emmeans::emmeans(fm8.11,specs = "Country")
#> NOTE: A nesting structure was detected in the fitted model:
#>     Province %in% Country
#> NOTE: Results may be misleading due to involvement in interactions
#>  Country     emmean    SE  df lower.CL upper.CL
#>  Australia     2.63 0.202 177     2.23     3.03
#>  Benin         3.34 0.403 177     2.55     4.14
#>  China         3.69 0.233 177     3.23     4.15
#>  Egypt         2.50 0.233 177     2.04     2.96
#>  Fiji          2.61 0.233 177     2.15     3.07
#>  Guam          2.34 0.403 177     1.55     3.14
#>  India         3.57 0.165 177     3.25     3.90
#>  Kenya         3.49 0.143 177     3.21     3.77
#>  Malaysia      4.03 0.134 177     3.77     4.30
#>  Mauritius     3.12 0.403 177     2.33     3.92
#>  PNG           3.65 0.403 177     2.85     4.45
#>  Philippines   3.61 0.233 177     3.15     4.07
#>  Puerto Rico   3.35 0.403 177     2.55     4.14
#>  Solomon IS    3.70 0.285 177     3.14     4.26
#>  Sri Lanks     3.23 0.233 177     2.78     3.69
#>  Thailand      3.84 0.202 177     3.44     4.24
#>  Vanuatu       2.76 0.403 177     1.97     3.56
#>  Vietnam       3.28 0.180 177     2.92     3.63
#> 
#> Results are averaged over the levels of: Province 
#> Confidence level used: 0.95