Example 8.2 from Experimental Design & Analysis for Tree Improvement

Exam8.2 presents the Diameter at breast height (Dbh) of 60 SeedLots under layout of row column design with 6 rows and 10 columns in 18 countries and 59 provinces of 18 selected countries.

References

Williams, E.R., Matheson, A.C. and Harwood, C.E. (2023).Experimental Design and Analysis for Tree Improvement. CSIRO Publishing.

See also

Author

Muhammad Yaseen (myaseen208@gmail.com)

Sami Ullah (samiullahuos@gmail.com)

Examples

data(DataExam8.2)
library(tidyverse)
library(ggplot2)

# Pg.

fm8.2  <-
  lmer(
    formula = dbhmean ~ Repl + Column + Contcompf + Contcompf:Standard +
              (1|Repl:Row ) + (1|Repl:Column ) + (1|Contcompf:Clone)
  , data    = DataExam8.2
    )
#> fixed-effect model matrix is rank deficient so dropping 5 columns / coefficients
summary(fm8.2)
#> Linear mixed model fit by REML. t-tests use Satterthwaite's method [
#> lmerModLmerTest]
#> Formula: dbhmean ~ Repl + Column + Contcompf + Contcompf:Standard + (1 |  
#>     Repl:Row) + (1 | Repl:Column) + (1 | Contcompf:Clone)
#>    Data: DataExam8.2
#> 
#> REML criterion at convergence: 746.9
#> 
#> Scaled residuals: 
#>      Min       1Q   Median       3Q      Max 
#> -2.23203 -0.64148  0.03063  0.52920  3.00825 
#> 
#> Random effects:
#>  Groups          Name        Variance Std.Dev.
#>  Contcompf:Clone (Intercept) 0.49500  0.7036  
#>  Repl:Row        (Intercept) 0.08023  0.2833  
#>  Repl:Column     (Intercept) 0.05286  0.2299  
#>  Residual                    0.39923  0.6318  
#> Number of obs: 300, groups:  Contcompf:Clone, 57; Repl:Row, 50; Repl:Column, 30
#> 
#> Fixed effects:
#>                          Estimate Std. Error       df t value Pr(>|t|)    
#> (Intercept)                5.9465     0.7910  61.1517   7.518 2.93e-10 ***
#> Repl2                      0.2958     0.2167  26.4669   1.365 0.183754    
#> Repl3                      0.3673     0.2167  26.4669   1.695 0.101836    
#> Repl4                      0.4848     0.2167  26.4669   2.237 0.033916 *  
#> Repl5                      0.5708     0.2167  26.4669   2.634 0.013923 *  
#> Column2                   -0.1497     0.2013  19.4810  -0.744 0.465921    
#> Column3                   -0.1278     0.2020  19.7650  -0.633 0.534107    
#> Column4                   -0.3884     0.2009  19.3185  -1.933 0.068042 .  
#> Column5                   -0.4594     0.2018  19.6615  -2.277 0.034139 *  
#> Column6                   -0.1114     0.2013  19.4681  -0.554 0.586148    
#> Contcompf1                 2.8234     0.7734  54.9050   3.651 0.000584 ***
#> Contcompf0:StandardUG323   2.8868     0.4362 214.4339   6.618 2.87e-10 ***
#> Contcompf0:StandardU6      0.4703     0.4251 202.8252   1.106 0.269914    
#> Contcompf0:StandardPN14    1.6185     0.4324 206.7230   3.743 0.000236 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Correlation matrix not shown by default, as p = 14 > 12.
#> Use print(x, correlation=TRUE)  or
#>     vcov(x)        if you need it
#> fit warnings:
#> fixed-effect model matrix is rank deficient so dropping 5 columns / coefficients
summary(fm8.2)$varcor
#>  Groups          Name        Std.Dev.
#>  Contcompf:Clone (Intercept) 0.70356 
#>  Repl:Row        (Intercept) 0.28326 
#>  Repl:Column     (Intercept) 0.22990 
#>  Residual                    0.63185 

anova(fm8.2)
#> Missing cells for: Contcompf0:Standard0, Contcompf1:StandardUG323, Contcompf1:StandardU6, Contcompf1:StandardPN14, Contcompf1:StandardSSOseed.  
#> Interpret type III hypotheses with care.
#> Type III Analysis of Variance Table with Satterthwaite's method
#>                     Sum Sq Mean Sq NumDF   DenDF F value    Pr(>F)    
#> Repl                3.2720  0.8180     4  26.467  2.0489 0.1162606    
#> Column              3.1018  0.6204     5  19.545  1.5539 0.2194719    
#> Contcompf           5.3203  5.3203     1  54.905 13.3265 0.0005845 ***
#> Contcompf:Standard 20.6587  6.8862     3 207.152 17.2488 4.896e-10 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

library(car)
Anova(fm8.2, type = "II", test.statistic = "Chisq")
#> Analysis of Deviance Table (Type II Wald chisquare tests)
#> 
#> Response: dbhmean
#>                      Chisq Df Pr(>Chisq)    
#> Repl                8.1957  4    0.08467 .  
#> Column              7.7694  5    0.16941    
#> Contcompf           4.6841  1    0.03044 *  
#> Contcompf:Standard 51.7463  3  3.392e-11 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1